revert parent - to water_heater
