+++ homeassistant/components/intouch/__init__.py
@@ -5,12 +5,10 @@ import voluptuous as vol
 from intouchclient import InTouchGateway
 
 from homeassistant.const import (
-    CONF_HOST, CONF_PASSWORD, CONF_USERNAME, EVENT_HOMEASSISTANT_START)
-from homeassistant.core import callback
+    CONF_HOST, CONF_PASSWORD, CONF_USERNAME)
 from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.discovery import async_load_platform
-from homeassistant.helpers.dispatcher import async_dispatcher_send
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -63,7 +61,7 @@ async def async_setup(hass, hass_config):
     hass.async_create_task(async_load_platform(
         hass, 'sensor', DOMAIN, {}, hass_config))
 
-    if len(heaters[0].rooms) > -1:  # TODO: should be > 0
+    if len(heaters[0].rooms) > -1:  # TODO: this is for testing, should be > 0
         hass.async_create_task(async_load_platform(
             hass, 'climate', DOMAIN, {}, hass_config))
